requirements.txt
streamlit==1.20.0
mysql-connector-python==8.0.32
pandas==1.5.3
pyodbc==4.0.35
Pillow==9.5.0
requests==2.31.0
